Added validation for testing set and fixed Makefile.
