+++ data/get_hashes.py
@@ -1,31 +1,29 @@
-import os
-import sys
-import json
+from __future__ import print_function, division
 
-from . import data
+import hashlib
 
-def get_all_hashes(data_dir):
-    """
-    Gets all files in all subdirectories of the supplied directory. 
-    Creates a dictionary where the paths to the files are the keys 
-    and the values are their corresponding hashes. 
-    
-    Returns file_hash, a dictionary of hashes. 
-    
-    Parameters
-    ----------
-    data_dir: path to a directory
-        The directory to search for files.
-    """
-    file_hashes = {}
-    for root, dirs, files in os.walk(data_dir):
-        for file_name in files:
-            file_path = os.path.join(root, file_name)
-            file_hashes[file_path] = data.generate_file_md5(file_path)
-    return file_hashes 
+# Get md5 hashes for `filename`. 
+def generate_file_md5(filename, blocksize=2**20):
+    m = hashlib.md5()
+    with open(filename, "rb") as f:
+        while True:
+            buf = f.read(blocksize)
+            if not buf:
+                break
+            m.update(buf)
+    return m.hexdigest()
 
-
-if __name__ == "__main__":
-    file_hashes = get_all_hashes('ds009')
-    with open('ds009_hashes.json', 'w') as out:
-        json.dump(file_hashes, out)
+# `d` is a dictionary in which the keys are file names and the items
+# are the hashes corresponding to the keys.
+# Checks whether the hash of each file key in `d` matches the 
+# corresponding hash item. 
+def check_hashes(d):
+    all_good = True
+    for k, v in d.items():
+        digest = generate_file_md5(k)
+        if v == digest:
+            print("The file {0} has the correct hash.".format(k))
+        else:
+            print("ERROR: The file {0} has the WRONG hash!".format(k))
+            all_good = False
+    return all_good
