+++ data/tests/test_data.py
@@ -8,8 +8,8 @@ try:
 except ImportError:
     from urllib2 import urlopen
 
-from .. import data
 from .. import get_hashes
+from .. import get_all_hashes
 
 
 def test_check_hashes():
@@ -18,9 +18,9 @@ def test_check_hashes():
         temp.flush()
         fname = temp.name
         d = {fname: "5b82f8bf4df2bfb0e66ccaa7306fd024"}
-        assert data.check_hashes(d)
+        assert get_hashes.check_hashes(d)
         d = {fname: "4b82f8bf4df2bfb0e66ccaa7306fd024"}
-        assert not data.check_hashes(d)
+        assert not get_hashes.check_hashes(d)
 
 def test_get_hashes(): 
     # Download class data to current directory. 
@@ -31,7 +31,7 @@ def test_get_hashes():
     output.write(ds107.read())
     output.close()
     # Now get all hashes of all files in current directory.
-    file_hashes = get_hashes.get_all_hashes('.')
+    file_hashes = get_all_hashes.get_all_hashes('.')
     # The class data should be in the dictionary of hashes, with the 
     # correct hash. 
     assert(file_hashes['./ds107_sub001_highres.nii'] == 'fd733636ae8abe8f0ffbfadedd23896c')
